Refs #22593: Add configuration example to windows ci
